added README, finished some commenting in SunGcrypt, started on change to dynamic link path on makefile

+++ SunGcrypt.h
@@ -30,7 +30,7 @@ public:
 	SunGcrypt();
 	bool CreateKey(string &key, size_t keyLength);
 	bool Encrypt(const string key, unsigned char* plainText, size_t plainTextLength, unsigned char* cipherText, unsigned int cipherTextLength);
-	bool Decrypt(const string key, unsigned char* cipherText, unsigned int cipherTextLength, unsigned char* plainText, size_t plainTextLength);
+	bool Decrypt(const string key, unsigned char* cipherText, unsigned int cipherTextLength, unsigned char* plainText, unsigned int plainTextLength);
 	bool AppendHMAC(const string key, unsigned char* data, unsigned int dataLength, unsigned char* signedData, unsigned int signedDataLength);
 	bool CheckHMAC(const string key, unsigned char* signedData, unsigned int signedDataLength, unsigned char* cipherText, unsigned int cipherTextLength);
 	unsigned int GetHMACLength();
